requirements.txt
appdirs==1.4.4
asgiref==3.3.4
certifi==2020.12.5
distlib==0.3.1
Django==3.2.3
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
filelock==3.0.12
pytz==2023.3
six==1.16.0
sqlparse==0.4.1
virtualenv==20.4.7
virtualenv-clone==0.5.4
